+++ cmd/shifu/main.go
@@ -36,7 +36,7 @@ func main() {
 			return
 		}
 
-		if err := server.Start(nil); err != nil {
+		if err := server.Start(nil, dir); err != nil {
 			slog.Error("Error starting Shifu", "error", err)
 		}
 	case "init":