+++ pkg/sass/sass_test.go
@@ -3,7 +3,6 @@ package sass
 import (
 	"context"
 	"github.com/emvi/shifu/pkg/cfg"
-	"github.com/emvi/shifu/pkg/storage"
 	"github.com/stretchr/testify/assert"
 	"os"
 	"path/filepath"
@@ -17,7 +16,6 @@ const (
 )
 
 func TestCompileSass(t *testing.T) {
-	fs := storage.NewFileStorage()
 	assert.NoError(t, os.RemoveAll(inDir))
 	assert.NoError(t, os.Mkdir(inDir, 0777))
 	in := filepath.Join(inDir, "test.scss")
@@ -28,7 +26,7 @@ func TestCompileSass(t *testing.T) {
 	cfg.Get().Sass.Entrypoint = "test.scss"
 	cfg.Get().Sass.Out = filepath.Join(outDir, "test.css")
 	cfg.Get().Sass.OutSourceMap = filepath.Join(outDir, "test.css.map")
-	Compile("", fs)
+	Compile("")
 	assert.FileExists(t, out)
 	assert.FileExists(t, cfg.Get().Sass.OutSourceMap)
 	content, err := os.ReadFile(out)
@@ -37,7 +35,6 @@ func TestCompileSass(t *testing.T) {
 }
 
 func TestWatchSass(t *testing.T) {
-	fs := storage.NewFileStorage()
 	assert.NoError(t, os.RemoveAll(inDir))
 	assert.NoError(t, os.Mkdir(inDir, 0777))
 	in := filepath.Join(inDir, "test.scss")
@@ -49,7 +46,7 @@ func TestWatchSass(t *testing.T) {
 	cfg.Get().Sass.Out = filepath.Join(outDir, "test.css")
 	cfg.Get().Sass.Watch = true
 	ctx, cancel := context.WithCancel(context.Background())
-	assert.NoError(t, Watch(ctx, "", fs))
+	assert.NoError(t, Watch(ctx, ""))
 	time.Sleep(time.Millisecond * 100)
 	assert.FileExists(t, out)
 	content, err := os.ReadFile(out)